% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmi_query.R
\name{fmi_query}
\alias{fmi_query}
\title{Construct a query to the FMI API}
\usage{
fmi_query(type = c("real-time", "daily", "monthly"), ...,
  api_key = getOption("fmi.api_key"))
}
\arguments{
\item{type}{a length 1 character vector specifying the measurement interval of the observations to request}

\item{...}{name-value pairs of length 1 character vectors, used as query parameters}

\item{api_key}{a length 1 character vector containing your personal FMI API key required to access the download service}
}
\value{
A length 1 character vector containing the URL for an FMI API query.
}
\description{
Construct a query to the FMI API
}
