% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fmi_set_key}
\alias{fmi_set_key}
\title{Set your FMI API key}
\usage{
fmi_set_key(api_key)
}
\arguments{
\item{api_key}{a length 1 character vector containing your personal FMI API key required to access the download service}
}
\description{
\code{fmi_set_key} saves your API key in your \code{options} for the duration of your R session so that you don't have to manually specify it each time you create a new query.
}
