% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmi_split_long_query.R
\name{fmi_split_long_query}
\alias{fmi_split_long_query}
\title{Split queries that exceed the maximum length for their type}
\usage{
fmi_split_long_query(query)
}
\arguments{
\item{query}{a character vector containing the URL used to request data from
the FMI API download service}
}
\value{
A vector of queries with long queries split into multiple smaller
ones.
}
\description{
Split queries that exceed the maximum length for their type
}
