% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmir.R
\docType{package}
\name{fmir-package}
\alias{fmir}
\alias{fmir-package}
\title{fmir: Access open weather data from the Finnish Meteorological
    Institute}
\description{
This package provides helper functions for
    easily accessing up to date, open weather data, made available by the
    Finnish Meteorological Institute. Use simple R syntax to create
    queries for the FMI API; download XML data from the API; and parse
    into a tidy R data frame. See
    https://en.ilmatieteenlaitos.fi/open-data for details.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mikmart/fmir}
}

}
\author{
\strong{Maintainer}: Mikko Marttila \email{mikmart@mikkomarttila.com}

}
\keyword{internal}
