% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{squash_list}
\alias{squash_list}
\title{Turn a list into a character vector}
\usage{
squash_list(x)
}
\arguments{
\item{x}{list}
}
\description{
Use to turn a (possibly nested) list to a character vector. Names for the
output vector are taken from the names of the top level list.
}
