% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmi_data.R
\name{fmi_data}
\alias{fmi_data}
\title{Download weather data from the FMI API}
\usage{
fmi_data(query)
}
\arguments{
\item{query}{a length 1 character vector containing the URL used to request data from the FMI API download service}
}
\value{
a \code{tbl_df} containing the requested data
}
\description{
Download weather data from the FMI API
}
