% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmi_query.R
\name{fmi_query}
\alias{fmi_query}
\title{Construct a query to the FMI API}
\usage{
fmi_query(type = c("real-time", "daily", "monthly"), ...)
}
\arguments{
\item{type}{A length 1 character vector specifying the measurement interval
of the observations to request}

\item{...}{Name-value pairs of character vectors, used as query parameters.
See details for possible values.}
}
\value{
A character vector containing query URLs for the FMI API.
}
\description{
Construct a query to the FMI API
}
\details{
The list of possible parameters passed in \code{...} depends on the type
and format of the query being constructed. Query-specific parameters are
fully documented in the \href{http://en.ilmatieteenlaitos.fi/open-data-manual-fmi-wfs-services}{FMI Open Data Manual}.

Common parameters include:

\describe{ \item{starttime, endtime}{A date or datetime specifying the
start/end of the interval to request data for. These must be in the
ISO-8601 format.} \item{place}{A string specifying the place of measurement
in general terms. E.g. \code{"Helsinki"}, \code{"Oulu"}} }
}
\examples{
fmi_query("real-time", place = "Helsinki")
}
\seealso{
\code{\link[=fmi_data]{fmi_data()}} to request data from the API.
}
