% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmi_set_key.R
\name{fmi_set_key}
\alias{fmi_set_key}
\alias{fmi_get_key}
\title{Set or get your FMI API key}
\usage{
fmi_set_key(api_key)

fmi_get_key()
}
\arguments{
\item{api_key}{A length 1 character vector containing your personal FMI API
key required to access the download service.}
}
\description{
Use \code{fmi_set_key()} to save your personal API key in \code{options} for the
duration of the R session so that it doesn't have to be manually specified
each time you create a new query. Alternatively, you can also set an
environment variable called \code{FMIR_API_KEY} in your \code{.Renviron} file to
persistently remember the API key. \code{fmi_get_key()} gets your API key by
checking the session option first (so you can change the key for your
session) and falls back to checking the environment variable.
}
\seealso{
\href{https://en.ilmatieteenlaitos.fi/open-data}{FMI Open Data
website} for obtaining a new API key.
}
